requirements.txt
playwright>=1.40.0
requests>=2.28.0
PyYAML>=6.0
Pillow>=9.0.0
pytesseract>=0.3.10
reportlab>=4.0
